requirements.txt
fastapi==0.118.0
uvicorn[standard]==0.37.0
python-multipart==0.0.6
langchain>=1.0.0
langchain-community>=0.4.0
langchain-groq>=0.1.0
langchain-openai>=0.1.0
langchain-chroma>=0.1.0
langchain-text-splitters>=1.0.0
chromadb>=1.3.0
PyPDF2==3.0.1
python-pptx==0.6.23
python-dotenv==1.0.0
pydantic==2.11.9
pydantic-settings>=2.12.0
celery==5.3.4
redis==5.0.1
flower==2.0.1
groq>=0.4.0
sqlalchemy>=2.0.0
openai>=1.0.0
tavily-python>=0.3.0
